Fix rental booking RPC date arithmetic for bigint span_days.
Production inserts failed because create_rental_booking_atomic used date + bigint, which Postgres rejects; cast the offset to integer so bookings can save.

+++ supabase/migrations/20260715190000_atomic_booking_creation.sql
@@ -1,5 +1,4 @@
 -- Phase 2: make availability validation and booking persistence one transaction.
--- This migration is intentionally forward-only and is not applied by this change.
 
 alter table public.bookings
   add column if not exists idempotency_key text null;
@@ -70,7 +69,9 @@ begin
     from public.bookings b
     where b.status in ('pending', 'approved', 'blocked')
       and b.event_date <= v_event_date + (v_span_days - 1)
-      and b.event_date + (greatest(coalesce(b.span_days, 1), 1) - 1) >= v_event_date
+      -- span_days is bigint on bookings; date + bigint is invalid — cast to int.
+      and b.event_date
+        + ((greatest(coalesce(b.span_days, 1), 1) - 1)::integer) >= v_event_date
       and (
         b.rental_item in (select value->>'rental_item' from jsonb_array_elements(p_items))
         or exists (
